+++ src/algorithms/PVT/libs/nmea_printer.cc
@@ -20,6 +20,7 @@
  */
 
 #include "nmea_printer.h"
+#include "gnss_sdr_filesystem.h"
 #include "rtklib_solution.h"
 #include "rtklib_solver.h"
 #include <glog/logging.h>
@@ -30,27 +31,6 @@
 #include <iostream>  // for cout, cerr
 #include <termios.h>
 #include <utility>
-
-// clang-format off
-#if HAS_STD_FILESYSTEM
-#include <system_error>
-namespace errorlib = std;
-#if HAS_STD_FILESYSTEM_EXPERIMENTAL
-#include <experimental/filesystem>
-namespace fs = std::experimental::filesystem;
-#else
-#include <filesystem>
-namespace fs = std::filesystem;
-#endif
-#else
-#include <boost/filesystem/operations.hpp>   // for create_directories, exists
-#include <boost/filesystem/path.hpp>         // for path, operator<<
-#include <boost/filesystem/path_traits.hpp>  // for filesystem
-#include <boost/system/error_code.hpp>       // for error_code
-namespace fs = boost::filesystem;
-namespace errorlib = boost::system;
-#endif
-// clang-format on
 
 
 Nmea_Printer::Nmea_Printer(const std::string& filename, bool flag_nmea_output_file, bool flag_nmea_tty_port, std::string nmea_dump_devname, const std::string& base_path)