+++ src/algorithms/libs/gnss_sdr_create_directory.cc
@@ -16,29 +16,9 @@
  */
 
 #include "gnss_sdr_create_directory.h"
+#include "gnss_sdr_filesystem.h"
 #include <exception>  // for exception
 #include <fstream>    // for ofstream
-
-// clang-format off
-#if HAS_STD_FILESYSTEM
-#include <system_error>
-namespace errorlib = std;
-#if HAS_STD_FILESYSTEM_EXPERIMENTAL
-#include <experimental/filesystem>
-namespace fs = std::experimental::filesystem;
-#else
-#include <filesystem>
-namespace fs = std::filesystem;
-#endif
-#else
-#include <boost/filesystem/operations.hpp>   // for create_directories, exists
-#include <boost/filesystem/path.hpp>         // for path, operator<<
-#include <boost/filesystem/path_traits.hpp>  // for filesystem
-#include <boost/system/error_code.hpp>       // for error_code
-namespace fs = boost::filesystem;
-namespace errorlib = boost::system;
-#endif
-// clang-format on
 
 bool gnss_sdr_create_directory(const std::string& foldername)
 {