+++ src/algorithms/telemetry_decoder/libs/tlm_utils.cc
@@ -15,6 +15,7 @@
  */
 
 #include "tlm_utils.h"
+#include "gnss_sdr_filesystem.h"
 #include <matio.h>
 #include <array>
 #include <cstdint>
@@ -22,26 +23,6 @@
 #include <iostream>
 #include <vector>
 
-// clang-format off
-#if HAS_STD_FILESYSTEM
-#include <system_error>
-namespace errorlib = std;
-#if HAS_STD_FILESYSTEM_EXPERIMENTAL
-#include <experimental/filesystem>
-namespace fs = std::experimental::filesystem;
-#else
-#include <filesystem>
-namespace fs = std::filesystem;
-#endif
-#else
-#include <boost/filesystem/operations.hpp>   // for remove
-#include <boost/filesystem/path.hpp>         // for path, operator<<
-#include <boost/filesystem/path_traits.hpp>  // for filesystem
-#include <boost/system/error_code.hpp>       // for error_code
-namespace fs = boost::filesystem;
-namespace errorlib = boost::system;
-#endif
-// clang-format on
 
 int save_tlm_matfile(const std::string &dumpfile)
 {